Added a first step toward auto-dependencies. Libs and apps now figure out what .h files they depend on and cache that info with the binaries.
find-lib.sh is now quieter, as the info is cached in the
appropriate .depend file (it also gets run less often).

I'll add dependence on the libraries (for apps) after the
break. I know how to do it, but I'd want to test it more
than I have time for. I'll archive how here, so I can see
it later, and some industrious person could do it while I'm
away:

   add
     $(APP_EXEC): $(join $(patsubst -L,,$(USES_LIBS_LD)),$(addprefix /lib,$(addsuffix .a,$(USES_LIBS_LIB))))
   to the .depend file

or something _real_ close to that.

NOTE!!!!

   This will work on coons right now, and other SGIS after about 4am.
I had to update gmake to get this working.
   This will work on cygwin right now. It already had the new gmake.