Comment improvements.
Re-work of command line arguments to label those that are clearly
obsolete, and make default values that match those of normal use.

Eliminates the need for the -grid parameter. Stream files and static files
will automatically resize the grid.

Adds a phantom control window, to allow use of the phantom button or the
button box.

Work in progress on light mode and loading static files after starting the
program - both do no harm.