Added command-line switch "-baseport num".
Using this option, it's now possible to run more than one instance of
nano at the same time on a multi-head machine like evans.
(Previously, if one user was running, another user could not run on a
different head.)

nano opens 5 or 6 ports.  They are used for things like collaboration
and remote rendering.  The problem was that the port numbers were
hardcoded, so the second instance of nano would fail to open the
ports.  There is a structure in microscape.c called WellKnownPorts
that holds the port numbers, and now there is a global varaible of that
type called wellKnownPorts.

The ports are now allocated in a contituous block, starting at NUM.
(They didn't used to be contiguous.)  If -baseport is not given on the
command-line, the first port number is 4501.

All of the code used in the nano application used WellKnownPorts to
get the port numbers, except for the genetic texture stuff.  There is
now a WellKnownPorts::remote_gaEngine, and the genetic texture library
has been modified.  But, it still needs to have a command-line option
added to the client.