Alexandra: Four probe measurments. The Keithley 2400 meter will do a four-probe measurement with an additional command that Alexandra identified. We also re-enabled the current source capability in the interface, and this seems to work correctly with four-probe measurements. Neil Snyder says that this four-probe measurement will be sufficient in the near term.
Connect to multiple instruments through the GPIB bus - this needs more
testing because we aren't using the second volt-meter at this time.

Aron:
Added a clear curves button which removes old curves from the graph. It
does not affect how curves are saved - that will be added later.

Makes sure the source, voltage or current, is graphed on the X-axis.