Added handlers to tie the mutex into the user interface. This code seems to work single-user:  as long as you don't open up the collaboration window and hit "release mutex", you don't even know it's there.  When you do that, nothing you do gets sent to the microscope until you hit "request mutex" again.
We need to do some rethinking of the user interface to make this entirely
correct:  for instance, if I release the mutex, then change the setpoint,
we don't get correct behavior.  The current implementation is that we display
whatever the user typed in, knowing that if the update was discarded the
incorrect value will be overwritten in a moment when we get the correct
value back from the microscope.  However, with the mutex released we never
send the message, so the incorrect value (that was thrown away) is never
cleared from the user interface.

Probably the easiest change to make would be to send a "set" command
immediately followed by a "query" command.  I believe that query commands
exist for MOST of the microscope's parameters, but don't know if the coverage
is complete.