Mutual exclusion is ready for live testing!?
There's one misfeature still in the user interface - we need to register
callbacks on the mutex before we construct the mutex, since they need to be
triggered by stuff in the constructor;  one possible solution would be to
defer the microscope initialization, which we're talking about doing for
other reasons.  Fault tolerance isn't working right, either, so if two people
are working and one quits while holding the lock the other is locked out.