Initializes the freespace_normal_force for directz to a bogus value at startup and when we go back into image mode. Checks for bogus value and doesn't apply force. freespace_normal_force will get initialized correctly when we get the InDirectZControl message.
Fixed a segfault when we create two flatten planes with the same name. The
updateflatten callback was not being removed, because the userdata that
was passed to remove_callback was different than that passed to
add_callback.

Removed state.relaxComp variable because it's obsolete and confusting.
Can't compile MicroscopeRcv.C anymore as a result.