Many changes to allow nano to compile and link using VC++ compiler and linker. One issue remains to be resolved: reading directory contents is not compatible with cygwin, so the colormap list is not filled.
Removed the BCDebug class.

Fixes for rotate-able scan direction. Nano now knows that the scan
angle has changed, and does something to compensate for it (rotates
the hand coordinates about the center of the scan area). We still have
to test if this is the right thing to do.

Changes to allow reading a stream file from the File menu. Microscope
initialization is encapsulated in the microscape.c:
CreateNewMicroscope function, which is called with an edited
inputstate structure if the user opens a different stream file.
Remaining issue: flat planes are stored in a static list variable,
which will be wrong if a stream file is loaded after a flat plane is
created.

Contour lines now display correctly running on the PC.