Various fixes, features:
This adds slow-line 3d functionality as described by Adam's specification
which can be found on the newest bug & features list.  Warning: in order
to selet your points for your line, you have to add them in *reverse*
order.  This was done to prevent the tip from crashing into whatever
object is between the two points (when the commit button is pressed,
the tip of the microscope goes to whatever the first point in the list is.
If there's an object between the two points....).

This makes the scanline fit the curvature of the surface a bit better than
it did before.  There were serious problems when a dip was present in the
surface on one side of the sample and the other side was relatively flat
with the rest of the surface.

This hides the scan-line when a screenshot is being saved so that it
doesn't show up in the final image saved to disk.