oops, forgot to commit this file when I removed getDisplayIndexForUser. Removes dependency on this now (again) non-existant function.  Instead, the vlib (vogl) display index is assumed to be zero.
In order to avoid confusion (yeah right;), the user display index is not
to be confused with the DISPLAY environment variable.  Vlib is designed to
allow multiple users.  This is the "handle" that is used to select a
particular user.  We've assumed all over nano.exe that the index is zero.