Unless NM_VERBOSE_MAKEFILE is set, compile and link lines are reduced to something that easily fits on a single line.  If you set NM_VERBOSE_MAKEFILE, then you see the while compile/link line.  The signal-to-noise is now reduced to the point that the compiler output is actually useful.
The easiest way to set the verbose mode is to do a setenv or do

   make NM_VERBOSE_MAKEFILE=1

I haven't tested this with VC++ since I don't have aron's setup to test
on.  Aron, there's a rule that has a comment which may mess up the
ifdef...else...endif stuff.  If it doesn't work, remove the comment or
figure out how to do comments correctly; I don't know what the right way
is.