Added the tracker server and client for the Magellan Puck. It can now grab the surface when you press the "*" button. It will grab the surface when the Phantom is in grab, light, scale, or touch stored mode. The transforms on this were annoying, but we can probably use it as an example to make the Navigate window act in screen space.
Slow line mode is no longer limited to the last two points you
specified: it starts at the first marker, and turns corners
appropriately to get to the last marker. BUG doesn't seem to work with
sweep style.

New light icon - a lamp that "orbits" the surface as the phantom
orientation changes, to show the direction the light is coming from.

Some pragmas for VC++ compiler to eliminate bogus conversion
warnings.

Made the measure line DZ measurements stay in real-world units when
the z-scale changes. So nmb_Line:: getIntercept returns a z value in
real-world units.

For some reason creation of the rulergrid texture was not symmetric -
in y direction, specifying 100% width only filled 50% of the
space. Fixed.

Tweaks to interaction.c

Changed "-verbosity" argument so it could also be "-verbose"

Added underlines to all menu items so they can be accessed with
keyboard shortcuts.

Changed the "Preferences" dialog to be the "Display Settings"
dialog. I decided not to implement preferences because it will only
cause confusion in a multi-user environment.

Disabled modifications with oscillating-sewing and
oscillating-forcecurve in the tcl interface.

Some more copyright notices.