Two-line change prevents the start-stop bug, at least by the means I was able to produce it (make a modification, then immediately press-and-release phantom button after letting up on first modification).  This is a spot-fix, not a general fix that would solve other problems in the future.
Proposed fix:  We've been talking about adding a function to Tclvars so
that we can differentiate between set-from-c-without-callbacks and
set-from-c-with-callbacks.  My 15-minute view is that we don't have to
have two different functions, because generally each variable will always
use one or the other.  So this switch (does this tclvar trigger callbacks
when assigned a value in C?) can be set only once, when the variable is
declared.

I'm not sure the proposed fix is prettier than the fine-grained per-operation
fix we've been talking about, but I think it's easier to understand/use,
easier to implement, and has fewer external repercussions (no hacking VRPN
to carry this bit in a side channel).

2001-04-25  Tom Hudson  <hudson@cs.unc.edu>

	* nmm_Microscope.C (RcvScanning) : add tweak to
        state.scanning.d_ignoreChange to avoid start-stop infinite loop
        problem.  This is a patch that should be corrected by systemic
        changes, since this could arise for other variables (although
        no others are quite so fragile?)