* openGL.c: Make sure that Texture Blending is only attempted on SGI's.  This fixes the crash bug that Adam found.
* nmg_GraphicsImpl.C: I had accidentally removed the updating of
the internal storage of the size of the plane in
causeGridRebuild.  This resulted in a situation where the surface
was being completely rebuilt every frame.  This was the culprit of
the severe speed slow down.

* nmg_SurfaceRegion.C: Use parent rederive function

* nmg_SurfaceMask.C (quadMasked): Made this a bit more efficient

* nmg_Surface.Ch (rederive): New function that regions call when
they realize that they need to rederive their mask plane.  This is
done through a call to the parent Surface class to ensure that any
other regions that would get affected by the call are in fact
updated.  This fixes a bug with stream files not updating
correctly.

* graphics_globals.ch: Removed global state guardian object

* nmg_StateGuardian.ch: Classes removed as I created them
hurriedly and they are currently unneccessary. (Although I believe
something like them would be worth while)