Bugfix to allow depth mesh rendering at a resolution higher than the resolution of the dataset.  More data!
There's still a memory leak in spm_gl.c::init_vertexArray for remote
rendering, but it occurs at startup, not ongoing, so doesn't need to be
fixed.  This leak is in the mainline code but should never be activated
(unless somebody else wrote code that calls init_vertexArray multiple
times).

2001-07-03  Tom Hudson  <hudson@cs.unc.edu>

	* nmg_GraphicsImpl.C (checkGridSize) : New function.
	(mainloop) : Call checkGridSize so derived classes can override.
	* nmg_RenderClientImpl.C (checkGridSize) : New function.