Comment out restore of g_just_color - spent an hour and a half tracking down a fix that was already in the main branch.
Add lots of spm_graphics_verbosity diagnostics used to hunt down the
above.  It's hard to keep a consistent, useful set of verbosity levels;
the easiest way without complicating the code might be to start
maintaining a levelization diagram of the system and use that to
assign verbosity levels to each object or function.  (Yes, Tom's
mania for drawing programs is surfacing again.)