Added support for -ptime <n>, which makes the simulator delay for <n> milliseconds on taking each point result.  Simulates real microscopes more accurately, although unfortunately that makes feelahead feel a little less desirable.
With a 5x5 point grid, 35 point samples are taken for feelahead,
meaning that with a 5 ms ptime (small but reasonable value) it takes
175 ms to gather a grid.  Comfortable movement speed means that the
user will get off the grid in under 175 ms unless the grid is made
quite large, in which case you lose resolution.  This is probably
another nice leverage point for adaptation:  measure how fast the
user is moving, as well as what the network latency is, and vary
the grid size proportionally.